refactor: adjusting project folders
