+++ node/.env
@@ -1,4 +1,4 @@
 DATABASE_NAME="dialog_rails"
 DATABASE_USERNAME="root"
-DATABASE_PASSWORD="dev@123"
+DATABASE_PASSWORD=""
 DATABASE_HOST="localhost"
