+++ node/src/server.js
@@ -3,7 +3,7 @@ const cors = require("cors");
 const routes = require('./routes');
 
 const app = express();
-const PORT = 8080;
+const PORT = 8000;
 
 app.use(express.json());
 app.use('/api', routes);
